Moves the static control flow analysis from the CLI tool into the lib.
